requirements.txt
# requirements.txt
fastapi
uvicorn
monai
torch
torchvision
opencv-python-headless
pillow
numpy
scikit-learn
python-multipart
